Split utils option helpers
