+++ tensordict/utils.py
@@ -57,7 +57,6 @@ from torch.nn.parameter import (
     UninitializedParameter,
     UninitializedTensorMixin,
 )
-from torch.utils._contextlib import _DecoratorContextManager
 from torch.utils.data._utils.worker import _generate_state
 
 try:
@@ -1807,370 +1806,18 @@ def _getitem_batch_size(batch_size, index):
     return torch.Size(out)
 
 
-# Repr / print options
-_REPR_OPTIONS = {
-    "show_batch_size": True,
-    "show_device": True,
-    "show_is_shared": True,
-    "show_shape": True,
-    "show_field_device": True,
-    "show_dtype": True,
-    "show_field_is_shared": True,
-    "show_grad": False,
-    "show_is_contiguous": False,
-    "show_is_view": False,
-    "show_storage_size": False,
-    "plain": False,
-    "sort_keys": "alphabetical",
-}
-
-_REPR_OPTIONS_KEYS = frozenset(_REPR_OPTIONS)
-
-
-_VERBOSE_FALSE_OVERRIDES = {
-    "show_device": False,
-    "show_is_shared": False,
-    "show_field_device": False,
-    "show_dtype": False,
-    "show_field_is_shared": False,
-}
-
-
-class set_printoptions(_DecoratorContextManager):
-    """Controls which attributes appear in TensorDict's ``__repr__`` output.
-
-    Can be used as a global setter (via :meth:`set`), a context manager, or a
-    decorator.  Follows the same pattern as :class:`set_lazy_legacy`.
-
-    Keyword Args:
-        verbose (bool, optional): Shortcut to control multiple options at once.
-            When ``False``, disables ``show_device``, ``show_is_shared``,
-            ``show_field_device``, ``show_dtype``, and ``show_field_is_shared``,
-            leaving only shape information visible.  Explicit keyword arguments
-            take precedence over ``verbose``.  ``verbose=True`` (the default)
-            has no effect — individual options keep their current values.
-        show_batch_size (bool, optional): Show ``batch_size`` in TensorDict repr.
-            Defaults to ``True``.
-        show_device (bool, optional): Show ``device`` in TensorDict repr.
-            Defaults to ``True``.
-        show_is_shared (bool, optional): Show ``is_shared`` in TensorDict repr.
-            Defaults to ``True``.
-        show_shape (bool, optional): Show ``shape`` in per-tensor field descriptors.
-            Defaults to ``True``.
-        show_field_device (bool, optional): Show ``device`` in per-tensor field
-            descriptors. Defaults to ``True``.
-        show_dtype (bool, optional): Show ``dtype`` in per-tensor field
-            descriptors. Defaults to ``True``.
-        show_field_is_shared (bool, optional): Show ``is_shared`` in per-tensor
-            field descriptors. Defaults to ``True``.
-        show_grad (bool, optional): Show ``requires_grad`` in per-tensor field
-            descriptors. Defaults to ``False``.
-        show_is_contiguous (bool, optional): Show ``is_contiguous`` in per-tensor
-            field descriptors. Defaults to ``False``.
-        show_is_view (bool, optional): Show ``is_view`` in per-tensor field
-            descriptors. Defaults to ``False``.
-        show_storage_size (bool, optional): Show ``storage_size`` (in bytes) in
-            per-tensor field descriptors. Defaults to ``False``.
-        plain (bool, optional): When ``True``, include a short summary of the
-            actual tensor values in the field descriptors. Defaults to ``False``.
-        sort_keys (str or callable, optional): Controls the order of keys in the
-            repr.  ``"alphabetical"`` (default) sorts keys lexicographically.
-            ``"insertion"`` preserves the order in which keys were added.
-            A callable is passed as the ``key`` argument to :func:`sorted`.
-
-    Examples:
-        >>> import torch
-        >>> from tensordict import TensorDict, set_printoptions
-        >>> td = TensorDict({"x": torch.randn(3, 4)})
-        >>> # Global
-        >>> set_printoptions(show_device=False, show_is_shared=False).set()
-        >>> print(td)
-        >>> # Context manager
-        >>> with set_printoptions(show_dtype=False):
-        ...     print(td)
-        >>> # Decorator
-        >>> @set_printoptions(show_is_shared=False)
-        ... def my_func(td):
-        ...     print(td)
-        >>> # Compact shapes-only output
-        >>> with set_printoptions(verbose=False):
-        ...     print(td)
-        >>> # Shapes-only but keep dtype visible
-        >>> with set_printoptions(verbose=False, show_dtype=True):
-        ...     print(td)
-
-    """
-
-    def __init__(self, *, verbose: bool = True, **kwargs) -> None:
-        super().__init__()
-        unknown = set(kwargs) - _REPR_OPTIONS_KEYS
-        if unknown:
-            raise TypeError(
-                f"Unknown printoptions: {unknown}. Valid options: {sorted(_REPR_OPTIONS_KEYS)}"
-            )
-        if not verbose:
-            merged = dict(_VERBOSE_FALSE_OVERRIDES)
-            merged.update(kwargs)
-            kwargs = merged
-        self._kwargs = kwargs
-
-    def clone(self) -> set_printoptions:
-        return type(self)(**self._kwargs)
-
-    def __enter__(self) -> None:
-        self.set()
-
-    def set(self) -> None:
-        global _REPR_OPTIONS
-        self._old = dict(_REPR_OPTIONS)
-        _REPR_OPTIONS.update(self._kwargs)
-
-    def __exit__(self, exc_type: Any, exc_value: Any, traceback: Any) -> None:
-        global _REPR_OPTIONS
-        _REPR_OPTIONS.update(self._old)
-
-
-def get_printoptions() -> dict:
-    """Returns the current TensorDict print options as a dict.
-
-    See :class:`set_printoptions` for details on each option.
-    """
-    return dict(_REPR_OPTIONS)
-
-
-# Lazy classes control (legacy feature)
-_DEFAULT_LAZY_OP = False
-_LAZY_OP = os.environ.get("LAZY_LEGACY_OP")
-
-
-class set_lazy_legacy(_DecoratorContextManager):
-    """Sets the behaviour of some methods to a lazy transform.
-
-    These methods include :meth:`~tensordict.TensorDict.view`, :meth:`~tensordict.TensorDict.permute`,
-    :meth:`~tensordict.TensorDict.transpose`, :meth:`~tensordict.TensorDict.squeeze`
-    and :meth:`~tensordict.TensorDict.unsqueeze`.
-
-    This property is dynamic, ie. it can be changed during the code execution, but
-    it won't propagate to sub-processes unless it has been called before the process
-    has been created.
-
-    """
-
-    def __init__(self, mode: bool) -> None:
-        super().__init__()
-        self.mode = mode
-
-    def clone(self) -> set_lazy_legacy:
-        # override this method if your children class takes __init__ parameters
-        return type(self)(self.mode)
-
-    def __enter__(self) -> None:
-        self.set()
-
-    def set(self) -> None:
-        global _LAZY_OP
-        self._old_mode = _LAZY_OP
-        _LAZY_OP = bool(self.mode)
-        # we do this such that sub-processes see the same lazy op than the main one
-        os.environ["LAZY_LEGACY_OP"] = str(_LAZY_OP)
-
-    def __exit__(self, exc_type: Any, exc_value: Any, traceback: Any) -> None:
-        global _LAZY_OP
-        _LAZY_OP = self._old_mode
-        os.environ["LAZY_LEGACY_OP"] = str(_LAZY_OP)
-
-
-def lazy_legacy(allow_none=False):
-    """Returns `True` if lazy representations will be used for selected methods."""
-    global _LAZY_OP
-    if _LAZY_OP is None and allow_none:
-        return None
-    elif _LAZY_OP is None:
-        return _DEFAULT_LAZY_OP
-    return strtobool(_LAZY_OP) if isinstance(_LAZY_OP, str) else _LAZY_OP
-
-
-def _legacy_lazy(func):
-    if not func.__name__.startswith("_legacy_"):
-        raise NameError(
-            f"The function name {func.__name__} must start with _legacy_ if it's decorated with _legacy_lazy."
-        )
-    func.LEGACY = True
-    return func
-
-
-# non tensor stack control
-_DEFAULT_CAPTURE_NONTENSOR_STACK = False
-_CAPTURE_NONTENSOR_STACK = os.environ.get("CAPTURE_NONTENSOR_STACK")
-
-
-class set_capture_non_tensor_stack(_DecoratorContextManager):
-    """A context manager or decorator to control whether identical non-tensor data should be stacked into a single NonTensorData object or a NonTensorStack.
-
-    Args:
-        mode (bool): Whether to capture non-tensor stacks. If ``False``, identical
-            non-tensor data will be stacked into a :class:`~tensordict.NonTensorStack`. If ``True``,
-            a single :class:`~tensordict.NonTensorData` object will contain the unique value, but with the desired batch-size.
-            Defaults to ``True``.
-
-    .. note:: Since v0.9, `capture_non_tensor_stack()` returns `False` by default.
-        You can set the value of :func:`~tensordict.capture_non_tensor_stack` through:
-
-        - The ``CAPTURE_NON_TENSOR_STACK`` environment variable;
-        - By setting ``set_capture_non_tensor_stack(val: bool).set()`` at the beginning of your script;
-        - By using ``set_capture_non_tensor_stack(val: bool)`` as a context manager or a decorator.
-
-        It is recommended to use the `set_capture_non_tensor_stack(False)` behavior.
-
-    .. seealso:: :class:`~tensordict.capture_non_tensor_stack`
-
-    Examples:
-        >>> with set_capture_non_tensor_stack(False):
-        ...     torch.stack([NonTensorData("a"), NonTensorData("a")])
-        NonTensorData("a", batch_size=[2])
-        >>> @set_capture_non_tensor_stack(False)
-        ... def my_function():
-        ...     return torch.stack([NonTensorData("a"), NonTensorData("a")])
-        >>> my_function()
-        NonTensorStack(["a", "a"], stack_dim=0)
-    """
-
-    def __init__(self, mode: bool) -> None:
-        super().__init__()
-        self.mode = mode
-
-    def clone(self) -> set_capture_non_tensor_stack:
-        # override this method if your children class takes __init__ parameters
-        return type(self)(self.mode)
-
-    def __enter__(self) -> None:
-        self.set()
-
-    def set(self) -> None:
-        global _CAPTURE_NONTENSOR_STACK
-        self._old_mode = _CAPTURE_NONTENSOR_STACK
-        _CAPTURE_NONTENSOR_STACK = bool(self.mode)
-        # we do this such that sub-processes see the same lazy op than the main one
-        os.environ["CAPTURE_NONTENSOR_STACK"] = str(_CAPTURE_NONTENSOR_STACK)
-
-    def __exit__(self, exc_type: Any, exc_value: Any, traceback: Any) -> None:
-        global _CAPTURE_NONTENSOR_STACK
-        _CAPTURE_NONTENSOR_STACK = self._old_mode
-        os.environ["CAPTURE_NONTENSOR_STACK"] = str(_CAPTURE_NONTENSOR_STACK)
-
-
-def capture_non_tensor_stack(allow_none=False):
-    """Get the current setting for capturing non-tensor stacks.
-
-    Args:
-        allow_none (bool, optional): If ``True``, returns ``None`` if no setting has been
-            specified. Otherwise, returns the default setting. Defaults to ``False``.
-
-    seealso: :func:`~tensordict.set_capture_non_tensor_stack`
-
-    Returns:
-        bool or None: The current setting for capturing non-tensor stacks.
-
-    """
-    global _CAPTURE_NONTENSOR_STACK
-    if _CAPTURE_NONTENSOR_STACK is None and allow_none:
-        return None
-    elif _CAPTURE_NONTENSOR_STACK is None:
-        return _DEFAULT_CAPTURE_NONTENSOR_STACK
-    elif (
-        isinstance(_CAPTURE_NONTENSOR_STACK, str)
-        and _CAPTURE_NONTENSOR_STACK.lower() == "none"
-    ):
-        return _DEFAULT_CAPTURE_NONTENSOR_STACK
-    return (
-        strtobool(_CAPTURE_NONTENSOR_STACK)
-        if isinstance(_CAPTURE_NONTENSOR_STACK, str)
-        else _CAPTURE_NONTENSOR_STACK
-    )
-
-
-# list to stack constrol
-_DEFAULT_LIST_TO_STACK = "1"
-_LIST_TO_STACK = os.environ.get("LIST_TO_STACK")
-
-
-class set_list_to_stack(_DecoratorContextManager):
-    """Context manager and decorator to control the behavior of list handling in TensorDict.
-
-    When enabled, lists assigned to a TensorDict will be automatically stacked along the batch dimension.
-    This can be useful for ensuring that lists of tensors or other elements are treated as stackable entities
-    within a TensorDict.
-
-    Current Behavior:
-        If a list is assigned to a TensorDict without this context manager, it will be converted to a numpy array
-            and wrapped in a NonTensorData if it cannot be cast to a Tensor.
-
-    Args:
-        mode (bool): If True, enables list-to-stack conversion. If False, disables it.
-
-    Example:
-        >>> with set_list_to_stack(True):
-        ...     td = TensorDict(a=[torch.zeros(()), torch.ones(())], batch_size=2)
-        ...     assert (td["a"] == torch.tensor([0, 1])).all()
-        ...     assert td[0]["a"] == 0
-        ...     assert td[1]["a"] == 1
-
-    .. seealso:: :func:`~tensordict.list_to_stack`.
-
-    """
-
-    def __init__(self, mode: bool) -> None:
-        super().__init__()
-        self.mode = mode
-
-    def clone(self) -> set_list_to_stack:
-        # override this method if your children class takes __init__ parameters
-        return type(self)(self.mode)
-
-    def __enter__(self) -> None:
-        self.set()
-
-    def set(self) -> None:
-        global _LIST_TO_STACK
-        self._old_mode = _LIST_TO_STACK
-        _LIST_TO_STACK = bool(self.mode)
-        os.environ["LIST_TO_STACK"] = str(_LIST_TO_STACK)
-
-    def __exit__(self, exc_type: Any, exc_value: Any, traceback: Any) -> None:
-        global _LIST_TO_STACK
-        _LIST_TO_STACK = self._old_mode
-        os.environ["LIST_TO_STACK"] = str(_LIST_TO_STACK)
-
-
-def list_to_stack(allow_none=False):
-    """Retrieves the current setting for list-to-stack conversion in TensorDict.
-
-    This function checks the global environment variable or the context manager setting to determine
-    whether lists should be automatically stacked when assigned to a TensorDict.
-
-    Current Behavior:
-        Returns the current setting for list-to-stack conversion. If the setting is not defined and `allow_none`
-            is True, it returns None. Otherwise, it returns the default setting.
-
-    Args:
-        allow_none (bool): If True, allows the function to return None if the setting is not defined.
-
-    Returns:
-        bool or None: The current setting for list-to-stack conversion.
-
-    .. seealso:: :class:`~tensordict.set_list_to_stack`.
-
-    """
-    global _LIST_TO_STACK
-    if _LIST_TO_STACK is None and allow_none:
-        return None
-    elif _LIST_TO_STACK is None:
-        return _DEFAULT_LIST_TO_STACK
-    elif isinstance(_LIST_TO_STACK, str) and _LIST_TO_STACK.lower() == "none":
-        return _DEFAULT_LIST_TO_STACK
-    return (
-        strtobool(_LIST_TO_STACK) if isinstance(_LIST_TO_STACK, str) else _LIST_TO_STACK
-    )
+from tensordict._utils_options import (  # noqa: F401
+    _legacy_lazy,
+    _REPR_OPTIONS,
+    capture_non_tensor_stack,
+    get_printoptions,
+    lazy_legacy,
+    list_to_stack,
+    set_capture_non_tensor_stack,
+    set_lazy_legacy,
+    set_list_to_stack,
+    set_printoptions,
+)
 
 
 def _convert_list_to_stack(
